chore: Fix clippy warnings and test compilation
FIX: Add missing dedup_safety_margin_hours in tests
- delivery_worker_test.rs: Added dedup_safety_margin_hours = 24
- queue_test.rs: Added dedup_safety_margin_hours = 24
- Fixes test compilation after Config struct update

FIX: Auto-fix clippy warnings
- Applied cargo clippy --fix for safe auto-fixes
- Fixed needless borrows and other lint suggestions
- 9 fixes in rest_api_test.rs

REMAINING: 107 warnings (mostly documentation and dead code in tests)
- 31 needless borrows (low priority, cosmetic)
- 6 missing documentation (crate-level, intentional)
- Test helper code warnings (acceptable for test code)

All critical warnings fixed. Tests compile and pass.

+++ shared/src/construct_server/pending_messages/storage.rs
@@ -78,7 +78,7 @@ impl PendingMessageStorage {
             .set(&key, &json_data)
             .ignore()
             // Set TTL
-            .expire(&key, PENDING_MSG_TTL_SECONDS as i64)
+            .expire(&key, PENDING_MSG_TTL_SECONDS)
             .ignore()
             // Add to index (sorted set by timestamp)
             .zadd(PENDING_INDEX_KEY, &data.temp_id, data.created_at)
